+++ src/cachelib/__init__.py
@@ -21,4 +21,4 @@ __all__ = [
     "MongoDbCache",
     "ValkeyCache",
 ]
-__version__ = "0.17.0.dev"
+__version__ = "0.16.1.dev"