+++ CHANGES.rst
@@ -4,7 +4,8 @@ Version 0.16.0
 unreleased
 
 - Change ``FileSystemCache`` default hash method to ``hashlib.sha256``
-  because ``hashlib.md5`` is not available in FIPS builds. :issue:`500`
+  because ``hashlib.md5`` is not available in FIPs builds. :issue:`500`
+- Ping backends servers on init to ensure they are reachable. :pr:`503`
 
 
 Version 0.15.4